backend/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.2
pytz==2020.1
sqlparse==0.3.1
